Update, refactor.
