+++ lib/tenji_minitest.rb
